// streampark-console/streampark-console-webapp-v2/src/enums/flinkEnum.ts
export enum ReleaseStateEnum {
    /** release failed */
    FAILED = -1,
    /** release done */
    DONE = 0,
    /** need release after modify task */
    NEED_RELEASE = 1,
    /** releasing */
    RELEASING = 2,
    /** release complete, need restart */
    NEED_RESTART = 3,
    /** need rollback */
    NEED_ROLLBACK = 4,
    /** project has changed, need to check the jar whether to be re-selected */
    NEED_CHECK = 5,
    /** revoked  */
    REVOKED = 10,
}

export enum OperationEnum {
    /** user operation: release app */
    RELEASE = 0,
    /** user operation: start app */
    START = 1,
    /** user operation: trigger savepoint */
    SAVEPOINT = 2,
    /** user operation: cancel app */
    CANCEL = 3,
}

export enum OptionStateEnum {
    /** Application which is currently action: none. */
    NONE = 0,
    /** Application which is currently action: deploying. */
    RELEASING = 1,
    /** Application which is currently action: cancelling. */
    CANCELLING = 2,
    /** Application which is currently action: starting. */
    STARTING = 3,
    /** Application which is currently action: savepointing. */
    SAVEPOINTING = 4,
}

export enum AppStateEnum {
    /** added new job to database  */
    ADDED = 0,
    /** The job has been received by the Dispatcher, and is waiting for the job manager to be created. */
    INITIALIZING = 1,
    /** Job is newly created, no task has started to run. */
    CREATED = 2,
    /** Application which is currently starting. */
    STARTING = 3,
    /** Application which is currently running. */
    RESTARTING = 4,
    /** Some tasks are scheduled or running, some may be pending, some may be finished. */
    RUNNING = 5,
    /** The job has failed and is currently waiting for the cleanup to complete. */
    FAILING = 6,
    /** The job has failed with a non-recoverable task failure.*/
    FAILED = 7,
    /** Job is being cancelled. */
    CANCELLING = 8,
    /** Job has been cancelled. */
    CANCELED = 9,
    /** All the job's tasks have successfully finished. */
    FINISHED = 10,
    /** The job has been suspended which means that it has been stopped but not been removed from a potential HA job store. */
    SUSPENDED = 11,
    /** The job is currently reconciling and waits for task execution report to recover state. */
    RECONCILING = 12,
    /** Lost */
    LOST = 13,
    /** MAPPING */
    MAPPING = 14,
    /** OTHER */
    OTHER = 15,
    /** has rollback */
    REVOKED = 16,
    /**
     * Lost track of flink job temporarily.
     * A complete loss of flink job tracking translates into LOST state.
     */
    SILENT = 17,
    /** Flink job has terminated vaguely, maybe FINISHED, CANCELED or FAILED */
    TERMINATED = 18,
    /** Flink job has terminated vaguely, maybe FINISHED, CANCELED or FAILED */
    POS_TERMINATED = 19,
    /** job SUCCEEDED on yarn */
    SUCCEEDED = 20,
    /** Job auto Health probe */
    KILLED = -9,
}